+++ 02-dtmoney/src/components/Dashboard/styles.ts
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+
+//Componente que renderiza o Dashboard
+export const Container = styled.main`
+    max-width: 1120px;
+    margin: 0 auto; /* Alinha o conteúdo do Dashboard no centro */
+    padding: 2.5rem 1rem;
+`;