+++ 02-dtmoney/src/styles/global.ts
@@ -4,6 +4,7 @@ export const GlobalStyle = createGlobalStyle`
     :root { /**  Variavel que define o estilo global */
         --background: #f0f2f5;
         --red: #E52E4D;
+        --green: #33CC95;
         --blue: #5429cc;
         --blue-light: #6933ff;
         --text-title: #363f5f;